get all loans api route and delete loan route functions
